pyproject.toml
[project]
name = "ptto115"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "p115client>=0.0.5.17.6",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
